\name{print.hct}
\alias{print.hct}
\alias{print}
\title{Print command for an hct object.
}
\description{Generic print command for hct object. Passes object to summary and then
prints summary.
}
\usage{
print(x,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{A hct object produced by hct.}
  \item{...}{Arguements passed to \code{summary.data.frame}}
}
\value{ Prints list created by \code{summary.hct}
}
\author{
David A. Schoenfeld}

\seealso{\code{\link{summary.hct}}}

\keyword{summary}% use one of  RShowDoc("KEYWORDS")
